[refactoring] will move navigationItems to main-navigation
